Action refactoring - split the run-in-renode.py
The run-in-renode.py script was splitted into smaller files
to improve code readability and facilitate later changes.

-action arguments are now converted to json
 before being passed to the script
-small typos fixes
-replaces sendcontrol('c') with sendline('') in some places
-fixes potential races with git repos sideloading
-adds 1 hour timeout to installing pip packages on target
